requirements.txt
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.1.0
langsmith>=0.1.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
crewai>=0.1.0
openai>=1.0.0